chadged graphicalMaterialScan -> graphicalScan; some changes in parameter definitions and order
