+++ UtilityApps/CMakeLists.txt
@@ -25,7 +25,7 @@ dd4hep_add_executable( print_materials src/print_materials.cpp USES DDRec )
 #-----------------------------------------------------------------------------------
 dd4hep_add_executable( materialScan    src/materialScan.cpp USES DDRec )
 #-----------------------------------------------------------------------------------
-dd4hep_add_executable( graphicalMaterialScan src/graphicalMaterialScan.cpp USES DDRec ROOT )
+dd4hep_add_executable( graphicalScan src/graphicalScan.cpp USES DDRec ROOT )
 #-----------------------------------------------------------------------------------
 #dd4hep_add_executable( pydd4hep     
 #  USES        [ROOT   REQUIRED COMPONENTS PyROOT]
